pipeline/requirements.txt
openai>=1.40.0
httpx>=0.27.0
pydantic>=2.7.0
tenacity>=8.2.3
tqdm>=4.66.0
rich>=13.7.0
cairosvg>=2.7.1
pillow>=10.0.0
pyyaml>=6.0.1
